update profile components structure.
